StructureRefactor
