+++ Demos/ClassesDemo/Program.cs
@@ -1,6 +1,6 @@
 ﻿namespace MSSA.Demos;
 
-class ClassesDemo
+class Program
 {
     static void Main(string[] args)
     {
