+++ Demos/DebugDemo/Program.cs
@@ -2,7 +2,7 @@
 
 namespace MSSA.Demos;
 
-class DebugDemo
+class Program
 {
     static void Main(string[] args)
     {
