+++ Demos/DoubleEndedQueueDemo/DoubleEndedQueueDemo.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "doubleendedqueuedemo", "doubleendedqueuedemo.csproj", "{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Debug|x64.Build.0 = Debug|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Debug|x86.Build.0 = Debug|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Release|x64.ActiveCfg = Release|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Release|x64.Build.0 = Release|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Release|x86.ActiveCfg = Release|Any CPU
+		{83AF29C3-99B7-4DE2-A3E7-02338686D5B1}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
