+++ Demos/LINQDemo/Program.cs
@@ -1,6 +1,6 @@
 ﻿namespace MSSA.Demos;
 
-class LINQDemo
+class Program
 {
     static void Main(string[] Args)
     {
