+++ Demos/SwitchCaseDemo/Program.cs
@@ -1,6 +1,6 @@
 ﻿namespace MSSA.Demos;
 
-class SwitchCaseDemo
+class Program
 {
     static void Main(string[] args)
     {
