+++ Demos/WhileLoopDemo/Program.cs
@@ -1,6 +1,6 @@
 ﻿namespace MSSA.Demos;
 
-class WhileLoopDemo
+class Program
 {
     static void Main(string[] args)
     {
